requirements.txt
langchain 
openai 
chromadb 
pypdf
tiktoken
gradio
wikipedia
streamlit
googletrans
google-search-results
python-dotenv
faiss-cpu
PyPDF2
unstructured
tabulate
pdf2image
streamlit_authenticator
google-api-python-client
streamlit-aggrid
pandas
deep-translator
aim